+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
+## [Unreleased]
+#### Added
+- `CardList` component
+
 ## [1.4.0-rc.3] - 2019-11-18
 #### Added
 - Google market img to `SuccessfulPayment` page